Make collisions work really well

+++ src/helpers/vec2f.h
@@ -4,16 +4,16 @@
 #include <cmath>
 
 struct Line {
-  Line(int x1_, int y1_, int x2_, int y2_) :
+  Line(double x1_, double y1_, double x2_, double y2_) :
       x1(x1_),
       y1(y1_),
       x2(x2_),
       y2(y2_) {};
   Line() = default;
-  int x1 = 0;
-  int y1 = 0;
-  int x2 = 0;
-  int y2 = 0;
+  double x1 = 0;
+  double y1 = 0;
+  double x2 = 0;
+  double y2 = 0;
 };
 
 class Vec2f {
